Work in progress for issue #2121
We now always set request_body_no_buffering. We rely on the Core Controller's own buffering instead.
The biggest issue I had to deal with so far was writing an output filter. This output filter
is installed when dealing with chunked request bodies, and re-chunks data sent to the Controller
(because Nginx dechunks bodies).

The following integration tests still fail:
Phusion Passenger for Nginx a Ruby app running on the root URI it should behave like an example web app buffers uploads

[ci:skip]